+++ src/commands/test.ts
@@ -3,7 +3,7 @@ import spinner from "../services/spinner";
 
 import { authenticate, BaseOptions, baseOptions, printError } from "../shared";
 import { validateMetadata } from "./deploy/handlers/validators";
-import handlers from "./snapshot/handlers";
+import handlers from "./test/handlers";
 
 export interface Options extends BaseOptions {
   config?: string;
@@ -11,8 +11,8 @@ export interface Options extends BaseOptions {
   service?: string;
 }
 
-export const command = "snapshot";
-export const desc = "Check all the alerts in the current service, create snapshots of the results, displays the results in the terminal, and outputs them to a file";
+export const command = "test";
+export const desc = "Test all the alerts in the current service, displays the results in the terminal, and outputs them to a file";
 
 export const builder: CommandBuilder<Options, Options> = (yargs) => {
   return yargs
@@ -20,18 +20,18 @@ export const builder: CommandBuilder<Options, Options> = (yargs) => {
       ...baseOptions,
       config: {
         type: "string",
-        desc: "The configuration folder of the service to snapshot. Defaults to the service specified in the .baselime folder, if it exists.",
+        desc: "The configuration folder of the service to test. Defaults to the service specified in the .baselime folder, if it exists.",
         alias: "c",
         default: ".baselime",
       },
       "out-file": { type: "string", desc: "The file to output the results to", alias: "o", default: "baselime-snapshot.json" },
-      service: { type: "string", desc: "The service to snapshot. This will be used to determine the service if no service is provided." },
+      service: { type: "string", desc: "The service to test. This will be used to determine the service if no service is provided." },
     })
     .example([
       [
         `
-      $0 snapshot
-      $0 snapshot --config .baselime --out-file file.json`,
+      $0 test
+      $0 test --config .baselime --out-file file.json`,
       ],
     ])
     .fail((message, err, yargs) => {
@@ -48,5 +48,5 @@ export async function handler(argv: Arguments<Options>) {
 
   service = service || (await validateMetadata(config!)).service;
 
-  await handlers.snapshot(format!, { service, outFile });
+  await handlers.test(format!, { service, outFile });
 }